"Downgrade composer plugin API version for compatibility"
This commit revises the version of the composer plugin API used in the project, downgrading from 2.6.0 to 2.3.0. The update was necessitated by the need for wider compatibility with stable plugins that have not yet been updated to use the newer API version. As such, this change should reduce the frequency of compatibility-related errors encountered during the use of said plugins. Future updates may be required when those plugins have been updated.